requirements.txt
async-timeout==5.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
humanize==4.12.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python-headless==4.7.0.72
piexif==1.1.3
Pillow==9.5.0
redis==5.2.1
requests==2.30.0
rq==2.3.2
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
yt-dlp
playwright
python-dotenv
polyline==1.4.0

